[#97] refactor: Mutation 함수 커스텀 훅으로 분리
