+++ src/components/Detail/ProductDetail/hooks/index.ts
@@ -1,0 +1,1 @@
+export * from '@/components/Detail/ProductDetail/hooks/useFavoriteMutation';
